tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.gradient.transparent.2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.gradient.transparent.2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.gradient.transparent.2</h1>
<p class="desc">Shadows are not drawn for transparent parts of gradient fills</p>


<script>
var t = async_test("Shadows are not drawn for transparent parts of gradient fills");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var gradient = ctx.createLinearGradient(0, 0, 100, 0);
  gradient.addColorStop(0, '#f00');
  gradient.addColorStop(0.499, '#f00');
  gradient.addColorStop(0.5, 'rgba(0,0,0,0)');
  gradient.addColorStop(1, 'rgba(0,0,0,0)');
  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 50, 50);
  ctx.fillStyle = '#0f0';
  ctx.fillRect(50, 0, 50, 50);
  ctx.shadowOffsetY = 50;
  ctx.shadowColor = '#0f0';
  ctx.fillStyle = gradient;
  ctx.fillRect(0, -50, 100, 50);

  _assertPixel(canvas, 25,25, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 75,25, 0,255,0,255);
  t.done();

});
</script>
